refactor: clean js monkey patching
