+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2021-2022 IRCAM – Centre Pompidou (France, Paris)
+Copyright (c) 2021-present IRCAM – Centre Pompidou (France, Paris)
 
 All rights reserved.
