+++ js/lib/errors.js
@@ -1,0 +1,8 @@
+class NotSupportedError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = 'NotSupportedError';
+  }
+}
+
+exports.NotSupportedError = NotSupportedError;
